+++ src/main/java/org/neoevolution/core/operator/reproduction/Reproduction.java
@@ -4,6 +4,6 @@ import org.neoevolution.core.model.Genotype;
 
 public interface Reproduction {
 
-    Genotype reproduce(Parents parents, int generation);
+    Genotype reproduce(Parents parents, Long generation);
 
 }
